Do not pass command line args to watch runner
We stop passing command line options from `runMocha` to `watchRun`. This
mechanism was only used to pass the `ui` parameter for resetting mocha
during watching. Since the command line options are duplicated in
`mocha.options` we get the `ui` parameter from there.

We also create the mocha instance in `runMocha` instead of passing it
together with the command line arguments. This means that `runMocha`
requires one argument instead of two.